Fixed a bunch of compiler warnings (MSVC) in the f2c code

+++ ext/f2c_libs/etime_.c
@@ -43,7 +43,7 @@ etime_(float *tarray)
 #endif
 	double t = clock();
 	tarray[1] = 0;
-	return tarray[0] = t / CLOCKS_PER_SECOND;
+	return (float) (tarray[0] = t / CLOCKS_PER_SECOND);
 #else
 	struct tms t;
